Moved add_list to io.xml.utils.py. Removed split_from_urn and split_unique_id and used parse_urn. Removed handlers file (redundant). Removed commented code.
